Enable improved POM support by default
This commit makes the experimental flag `IMPROVED_POM_SUPPORT` the default.
The flag is still there for backwards compatibility but has effectively no
impact. As a consequence, the behavior of improved POM support is now the
default, which implies that:

- Maven dependencies packaged as `pom` or `jar` now have derived variants
(`compile` and `runtime`) and we properly choose between the variants based
on the consumer attributes
- platform dependencies using the `platform` and `enforcedPlatform` keywords
are enabled

Enabling improved POM support by default is a **breaking change**: there's
a risk that resolved dependencies is different, in particular because we
will now only include the `compile` dependencies of a POM file whenever the
consumer asks for the API variant. There are also some changes in the
dependency insight reports due to the use of attribute based matching instead
of configuration selection.

Last but not least, this commit is likely to introduce a small performance
regression due to attribute based selection.